requirements.txt
torch
torchvision
hydra-core
tensorboard
numpy
omegaconf
Pillow